SI-4694 Removed extra vmem = NULL from param.h define. Changed the alignment of the param_t types to be on a 8-byte boundary since the new vaddr field is 64-bits. Changed the vmem_t to be 8-byte aligned because the vaddr is a 64-bit usigned integer. Extended the vmem_calc_crc32() be be able to calculate CRC32 on a length greater than 32-bits, now 64-bits. Changed the VMEM server version to be 3, with 64-bit support for the size field. The VMEM server now handles 64-bit lengths in both upload, download and crc32 operations.

+++ src/vmem/vmem_crc32.c
@@ -12,10 +12,10 @@
 #include "vmem/vmem_crc32.h"
 #include "csp/csp_crc32.h"
 
-uint32_t vmem_calc_crc32(uint64_t addr, intptr_t len, void * buffer, uint32_t buffer_len) {
+uint32_t vmem_calc_crc32(uint64_t addr, uint64_t len, void * buffer, uint32_t buffer_len) {
 
 	csp_crc32_t crc_obj;
-	uint32_t count = 0;
+	uint64_t count = 0;
 	uint32_t chunk_len;
 
 	csp_crc32_init(&crc_obj);
